fixed bug for starter code

+++ algorithms/starter_code.py
@@ -7,8 +7,9 @@ def data2dag(data, num_nodes):
   dag = snap.TNGraph.New()
   for i in range(num_nodes):
     dag.AddNode(i)
+
   for i in range(data.shape[0]):
-    dag.AddEdge(data[i][0], data[i][1])
+    dag.AddEdge(int(data[i][0]), int(data[i][1]))
   FOut = snap.TFOut("../data/youtube.graph")
   dag.Save(FOut)
   return dag
@@ -28,7 +29,7 @@ def main():
     social_network = snap.TNGraph.Load(FIn)
   else:
     edges = pd.read_csv("../data/edges.csv", sep='\t', index_col=0)
-    edges = np.asarray(edges)
+    edges = np.asarray(edges).astype(int)
     social_network = data2dag(edges, nodes.shape[0])
